Add Pal page to the app
This commit adds a new page to the app that displays the details of a pal.
The page is accessible by clicking on a pal card from the home page.
The page displays the pal's details, breeding list, and breeding card.
The breeding list is a list of breeding cards that display the pal's
breeding combinations. Each pal image in the breeding card leads to the
pal's details page.